Extract all functionality related to tracking the registration of services into a dedicated tracker
