+++ src/Autofac/Core/Registration/ComponentRegistry.cs
@@ -27,8 +27,6 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Diagnostics;
-using System.Globalization;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using Autofac.Builder;
@@ -54,34 +52,31 @@ namespace Autofac.Core.Registration
         /// </summary>
         private readonly object _synchRoot = new object();
 
-        /// <summary>
-        /// External registration sources.
-        /// </summary>
-        private readonly List<IRegistrationSource> _dynamicRegistrationSources = new List<IRegistrationSource>();
-
-        /// <summary>
-        /// All registrations.
-        /// </summary>
-        private readonly List<IComponentRegistration> _registrations = new List<IComponentRegistration>();
-
-        /// <summary>
-        /// Keeps track of the status of registered services.
-        /// </summary>
-        private readonly ConcurrentDictionary<Service, ServiceRegistrationInfo> _serviceInfo = new ConcurrentDictionary<Service, ServiceRegistrationInfo>();
-
         private readonly ConcurrentDictionary<IComponentRegistration, IEnumerable<IComponentRegistration>> _decorators
             = new ConcurrentDictionary<IComponentRegistration, IEnumerable<IComponentRegistration>>();
 
         private readonly IDictionary<string, object> _properties;
+        private readonly IRegisteredServicesTracker _registeredServicesTracker;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ComponentRegistry"/> class.
+        /// </summary>
+        /// <param name="registeredServicesTracker">The tracker for the registered services.</param>
+        /// <param name="properties">The properties used during component registration.</param>
+        internal ComponentRegistry(IRegisteredServicesTracker registeredServicesTracker, IDictionary<string, object> properties)
+        {
+            _properties = properties;
+            _registeredServicesTracker = registeredServicesTracker;
+            ((IComponentRegistryBuilder)this).Register(new SelfComponentRegistration());
+        }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ComponentRegistry"/> class.
         /// </summary>
         /// <param name="properties">The properties used during component registration.</param>
         internal ComponentRegistry(IDictionary<string, object> properties)
+            : this(new DefaultRegisteredServicesTracker(), properties)
         {
-            _properties = properties;
-            ((IComponentRegistryBuilder)this).Register(new SelfComponentRegistration());
         }
 
         /// <summary>
@@ -112,8 +107,7 @@ namespace Autofac.Core.Registration
         /// <param name="disposing"><c>true</c> to release both managed and unmanaged resources; <c>false</c> to release only unmanaged resources.</param>
         protected override void Dispose(bool disposing)
         {
-            foreach (var registration in _registrations)
-                registration.Dispose();
+            _registeredServicesTracker.Dispose();
 
             base.Dispose(disposing);
         }
@@ -126,22 +120,12 @@ namespace Autofac.Core.Registration
         /// <returns>True if a registration exists.</returns>
         bool IComponentRegistry.TryGetRegistration(Service service, out IComponentRegistration registration)
         {
-            if (service == null) throw new ArgumentNullException(nameof(service));
-
-            var info = GetInitializedServiceInfoOrDefault(service);
-            if (info != null && info.TryGetRegistration(out registration))
-                return true;
-
-            lock (_synchRoot)
-            {
-                info = GetInitializedServiceInfo(service);
-                return info.TryGetRegistration(out registration);
-            }
+            return _registeredServicesTracker.TryGetRegistration(service, out registration);
         }
 
         bool IComponentRegistryBuilder.IsRegistered(Service service)
         {
-            return IsRegistered(service);
+            return _registeredServicesTracker.IsRegistered(service);
         }
 
         /// <summary>
@@ -151,21 +135,7 @@ namespace Autofac.Core.Registration
         /// <returns>True if the service is registered.</returns>
         bool IComponentRegistry.IsRegistered(Service service)
         {
-            return IsRegistered(service);
-        }
-
-        private bool IsRegistered(Service service)
-        {
-            if (service == null) throw new ArgumentNullException(nameof(service));
-
-            var info = GetInitializedServiceInfoOrDefault(service);
-            if (info != null && info.IsRegistered)
-                return true;
-
-            lock (_synchRoot)
-            {
-                return GetInitializedServiceInfo(service).IsRegistered;
-            }
+            return _registeredServicesTracker.IsRegistered(service);
         }
 
         /// <summary>
@@ -187,54 +157,7 @@ namespace Autofac.Core.Registration
         {
             if (registration == null) throw new ArgumentNullException(nameof(registration));
 
-            lock (_synchRoot)
-            {
-                AddRegistration(registration, preserveDefaults);
-                UpdateInitialisedAdapters(registration);
-            }
-        }
-
-        private void UpdateInitialisedAdapters(IComponentRegistration registration)
-        {
-            var adapterServices = new List<Service>();
-            foreach (var serviceInfo in _serviceInfo)
-            {
-                if (serviceInfo.Value.ShouldRecalculateAdaptersOn(registration))
-                {
-                    adapterServices.Add(serviceInfo.Key);
-                }
-            }
-
-            if (adapterServices.Count == 0)
-                return;
-
-            Debug.WriteLine(
-                string.Format(
-                    CultureInfo.InvariantCulture,
-                    "[Autofac] Component '{0}' provides services that have already been adapted. Consider refactoring to ContainerBuilder.Build() rather than Update().",
-                    registration));
-
-            var adaptationSandbox = new AdaptationSandbox(
-                _dynamicRegistrationSources.Where(rs => rs.IsAdapterForIndividualComponents),
-                registration,
-                adapterServices);
-
-            // Adapter registrations come from sources, so they are added with originatedFromSource: true
-            var adapters = adaptationSandbox.GetAdapters();
-            foreach (var adapter in adapters)
-                AddRegistration(adapter, true, true);
-        }
-
-        protected virtual void AddRegistration(IComponentRegistration registration, bool preserveDefaults, bool originatedFromSource = false)
-        {
-            foreach (var service in registration.Services)
-            {
-                var info = GetServiceInfo(service);
-                info.AddImplementation(registration, preserveDefaults, originatedFromSource);
-            }
-
-            _registrations.Add(registration);
-
+            _registeredServicesTracker.AddRegistration(registration, preserveDefaults);
             GetRegistered()?.Invoke(this, new ComponentRegisteredEventArgs(this, registration));
         }
 
@@ -243,11 +166,7 @@ namespace Autofac.Core.Registration
         /// </summary>
         IEnumerable<IComponentRegistration> IComponentRegistry.Registrations
         {
-            get
-            {
-                lock (_synchRoot)
-                    return _registrations.ToArray();
-            }
+            get { return _registeredServicesTracker.Registrations; }
         }
 
         /// <summary>
@@ -259,17 +178,7 @@ namespace Autofac.Core.Registration
         /// <returns>Registrations supporting <paramref name="service"/>.</returns>
         IEnumerable<IComponentRegistration> IComponentRegistry.RegistrationsFor(Service service)
         {
-            if (service == null) throw new ArgumentNullException(nameof(service));
-
-            var info = GetInitializedServiceInfoOrDefault(service);
-            if (info != null)
-                return info.Implementations.ToArray();
-
-            lock (_synchRoot)
-            {
-                info = GetInitializedServiceInfo(service);
-                return info.Implementations.ToArray();
-            }
+            return _registeredServicesTracker.RegistrationsFor(service);
         }
 
         /// <inheritdoc />
@@ -286,7 +195,7 @@ namespace Autofac.Core.Registration
                     if (service is DecoratorService || !(service is IServiceWithType swt)) continue;
 
                     var decoratorService = new DecoratorService(swt.ServiceType);
-                    var decoratorRegistrations = ((IComponentRegistry)this).RegistrationsFor(decoratorService);
+                    var decoratorRegistrations = _registeredServicesTracker.RegistrationsFor(decoratorService);
                     result.AddRange(decoratorRegistrations);
                 }
 
@@ -304,7 +213,7 @@ namespace Autofac.Core.Registration
             {
                 lock (_synchRoot)
                 {
-                    foreach (IComponentRegistration registration in _registrations)
+                    foreach (IComponentRegistration registration in _registeredServicesTracker.Registrations)
                     {
                         value(this, new ComponentRegisteredEventArgs(this, registration));
                     }
@@ -328,17 +237,8 @@ namespace Autofac.Core.Registration
         /// <param name="source">The source to register.</param>
         void IComponentRegistryBuilder.AddRegistrationSource(IRegistrationSource source)
         {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-
-            lock (_synchRoot)
-            {
-                _dynamicRegistrationSources.Insert(0, source);
-                foreach (var serviceRegistrationInfo in _serviceInfo)
-                    serviceRegistrationInfo.Value.Include(source);
-
-                var handler = GetRegistrationSourceAdded();
-                handler?.Invoke(this, new RegistrationSourceAddedEventArgs(this, source));
-            }
+            _registeredServicesTracker.AddRegistrationSource(source);
+            GetRegistrationSourceAdded()?.Invoke(this, new RegistrationSourceAddedEventArgs(this, source));
         }
 
         /// <summary>
@@ -346,13 +246,7 @@ namespace Autofac.Core.Registration
         /// </summary>
         IEnumerable<IRegistrationSource> IComponentRegistry.Sources
         {
-            get
-            {
-                lock (_synchRoot)
-                {
-                    return _dynamicRegistrationSources.ToArray();
-                }
-            }
+            get { return _registeredServicesTracker.Sources; }
         }
 
         /// <summary>
@@ -373,7 +267,7 @@ namespace Autofac.Core.Registration
             {
                 lock (_synchRoot)
                 {
-                    foreach (IRegistrationSource source in _dynamicRegistrationSources)
+                    foreach (IRegistrationSource source in _registeredServicesTracker.Sources)
                     {
                         value(this, new RegistrationSourceAddedEventArgs(this, source));
                     }
@@ -389,61 +283,6 @@ namespace Autofac.Core.Registration
                     _properties[MetadataKeys.RegistrationSourceAddedPropertyKey] = GetRegistrationSourceAdded() - value;
                 }
             }
-        }
-
-        private ServiceRegistrationInfo GetInitializedServiceInfo(Service service)
-        {
-            var info = GetServiceInfo(service);
-            if (info.IsInitialized)
-                return info;
-
-            if (!info.IsInitializing)
-                info.BeginInitialization(_dynamicRegistrationSources);
-
-            while (info.HasSourcesToQuery)
-            {
-                var next = info.DequeueNextSource();
-                foreach (var provided in next.RegistrationsFor(service, s => ((IComponentRegistry)this).RegistrationsFor(s)))
-                {
-                    // This ensures that multiple services provided by the same
-                    // component share a single component (we don't re-query for them)
-                    foreach (var additionalService in provided.Services)
-                    {
-                        var additionalInfo = GetServiceInfo(additionalService);
-                        if (additionalInfo.IsInitialized || additionalInfo == info) continue;
-
-                        if (!additionalInfo.IsInitializing)
-                            additionalInfo.BeginInitialization(_dynamicRegistrationSources.Where(src => src != next));
-                        else
-                            additionalInfo.SkipSource(next);
-                    }
-
-                    AddRegistration(provided, true, true);
-                }
-            }
-
-            info.CompleteInitialization();
-            return info;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private ServiceRegistrationInfo GetServiceInfo(Service service)
-        {
-            if (_serviceInfo.TryGetValue(service, out var existing))
-                return existing;
-
-            var info = new ServiceRegistrationInfo(service);
-            _serviceInfo.TryAdd(service, info);
-            return info;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private ServiceRegistrationInfo GetInitializedServiceInfoOrDefault(Service service)
-        {
-            if (_serviceInfo.TryGetValue(service, out var existing) && existing.IsInitialized)
-                return existing;
-
-            return null;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
