+++ src/Autofac/Core/Registration/IRegisteredServicesTracker.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Autofac.Core.Registration
+{
+    internal interface IRegisteredServicesTracker : IDisposable
+    {
+        void AddRegistration(IComponentRegistration registration, bool preserveDefaults, bool originatedFromSource = false);
+
+        /// <summary>
+        /// Add a registration source that will provide registrations on-the-fly.
+        /// </summary>
+        /// <param name="source">The source to register.</param>
+        void AddRegistrationSource(IRegistrationSource source);
+
+        /// <summary>
+        /// Attempts to find a default registration for the specified service.
+        /// </summary>
+        /// <param name="service">The service to look up.</param>
+        /// <param name="registration">The default registration for the service.</param>
+        /// <returns>True if a registration exists.</returns>
+        bool TryGetRegistration(Service service, out IComponentRegistration registration);
+
+        /// <summary>
+        /// Gets the registered components.
+        /// </summary>
+        IEnumerable<IComponentRegistration> Registrations { get; }
+
+        /// <summary>
+        /// Gets the registration sources that are used by the registry.
+        /// </summary>
+        IEnumerable<IRegistrationSource> Sources { get; }
+
+        bool IsRegistered(Service service);
+
+        /// <summary>
+        /// Selects from the available registrations after ensuring that any
+        /// dynamic registration sources that may provide <paramref name="service"/>
+        /// have been invoked.
+        /// </summary>
+        /// <param name="service">The service for which registrations are sought.</param>
+        /// <returns>Registrations supporting <paramref name="service"/>.</returns>
+        IEnumerable<IComponentRegistration> RegistrationsFor(Service service);
+    }
+}
